requirements.txt
click==8.0.1
colorama==0.4.4
ghp-import==2.0.1
importlib-metadata==4.8.1
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
mergedeep==1.3.4
mkdocs==1.2.2
mkdocs-macros-plugin==0.6.0
mkdocs-material==7.2.7
mkdocs-material-extensions==1.0.3
mkdocs-simple-hooks==0.1.3
packaging==21.0
Pygments==2.10.0
pymdown-extensions==8.2
pyparsing==2.4.7
python-dateutil==2.8.2
PyYAML==5.4.1
pyyaml-env-tag==0.1
six==1.16.0
termcolor==1.1.0
watchdog==2.1.5
zipp==3.5.0
